Start working on Froyo::find_all
For "list", we need to print the names of froyodevs (there many be more
than one) found on the system. We can borrow some code from Melvin for
finding blockdevs with the right signature, but to construct a Froyo we
need to parse the json metadata, which isn't implemented yet.

Signed-off-by: Andy Grover <agrover@redhat.com>