Replace MDA timestamp/serial with timestamp/nanoseconds
Previously, one of the new MDA's values, the serial, was dependent on
the other MDA -- whether the timestamp was identical or not. This meant
that writing the same MDA across multiple disks could result in different
MDA values, and there was no way to tell which one was the newest.

Instead, use the identical time values for writes across all disks. On each
disk we need to look at the existing MDAs to determine which one to
overwrite (the older one) but we can compare time values across multiple
disks to determine which one has the most up-to-date metadata.

Signed-off-by: Andy Grover <agrover@redhat.com>