A new approach to ser/deserialization
Instead of writing implementations of Serialize and Deserialize for our
primary data structures create secondary "Save" structures that are easily
serializable that we can derive from the primary structures, and then use
to build the primary structures ourselves from the metadata later. This
will allow us to properly handle errors and construct references to
refcounted structs.

The only place where we still manually derive Ser and De is for Sector and
SectorOffset.

Signed-off-by: Andy Grover <agrover@redhat.com>