Combine LinearDevs for metadata & data
It's simpler to have a LinearDev keep track of both together, since they
will never be on different blockdevs. It's just one more pair of
offset/length to just have one struct.

This makes LinearDev::create() a bear, but other areas are simplified
because we just have LinearDev instead of (LinearDev, LinearDev)s
everywhere.

Furthermore, do not have a separate LinearDevSave vector in FroyoSave, but
have each RaidDevSave contain a vec of LinearDevSave instead of ids.

Signed-off-by: Andy Grover <agrover@redhat.com>