thin: Add support for status, and creating devnodes and mkfs.xfs
I think we want to not rely on udev to create devnodes. This avoids a
race when mkfs-ing on a device we just created.

Call out to mkfs.xfs to make an xfs filesystem on the new thin device.

Add a status() method that parses the thin dev's status message (pretty
simple, but we want to know how much of the device is allocated.)

Signed-off-by: Andy Grover <agrover@redhat.com>