Make things refcounted for DBus
Everything is going to be single-threaded, but all these method handlers
are closures that need to close over a Froyo (most of them) and Create
needs to modify the list of Froyos, to add the one it just created.

Therefore we end up with refcounting of each Froyo as well as the Vec.

Name and SetName work for froyos. We'll use the outer refcount when next
we hook up Create. Wheeeeeeeeeee!

Signed-off-by: Andy Grover <agrover@redhat.com>