Change list cmd to use DBus
We want to transition all commandline tools (except maybe dev subcommands)
to talk to the froyo service rather than building their own list of froyos.
This is the first one. We're still using an older dbus-rs API, message
instead of arg, because arg ties things more tightly with the type system
and this makes working with the retval from GetManagedObjects, which is
"a{oa{sa{sv}}}", a little too tricky for me at the moment.

Signed-off-by: Andy Grover <agrover@redhat.com>