Rework state machine
Integrate Reshape states into overall FroyoState as FroyoState::Good(
FroyoRunningState::Reshaping(ReshapeState)).

Change things named "*Status" to be named "*State", for consistency.

Add an additional Initializing state to FroyoState, to allow all raids to
resync before we try creating the thin volume. (Actually the xfs mkfs is
the thing that takes a long time on not-fully-synced raids.) This will let
create() succeed quickly, but then we still have to wait for the Froyodev
to be usable.

Make check_reshape() into a more general check_state, that handles calling
the reshape state machine, and also the Initializing->Good transition.

Add "initializing" bit in status, to indicate the raid is not yet online
due to initial resync.

Fix some space-related bugs.

Signed-off-by: Andy Grover <agrover@redhat.com>