Make BlockDev.linear_devs a btreemap and add RaidDev::destroy()
This is to make it possible to remove a lineardev more easily, which
happens when we recreate raids during reshape. Failure to do this means
we still think the space on the blockdev is in use.

Change recreate_empty_degraded_raids to call RaidDev::destroy() instead of
teardown. This removes the dm mapping and also frees the space, based on
the above change.

Signed-off-by: Andy Grover <agrover@redhat.com>